fix(stepper): correct step_mask handling and step pulse timing
Fixed step_mask bit checks to correctly map step bits per axis and prevent unintended axis stepping.
Added an early return when step_mask is zero to avoid unnecessary GPIO and timer operations.
Improved CC2 scheduling flow by disabling and clearing CC2 interrupt flags before updating compare values, reducing race-related timing issues.
Added a fallback to reset step pins immediately when the compare target is already missed, preventing overly long high pulses.

+++ drivers/grbl_stepper_controller.h
@@ -9,7 +9,7 @@
 int stepper_controller_set_period(const struct device *dev, uint32_t cycles);
 
 /* Set the Step bits high for the axes specified in the mask */
-void stepper_controller_set_steps(const struct device *dev, uint8_t step_mask);
+void stepper_controller_set_steps(const struct device *dev, uint16_t step_mask);
 
 void stepper_controller_reset_steps(const struct device *dev);
